% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOutput.R
\name{computeOccupancyStatistics}
\alias{computeOccupancyStatistics}
\title{Computes occupancy statistics}
\usage{
computeOccupancyStatistics(out, start = 1)
}
\arguments{
\item{out}{result of a simulation}
}
\value{
a list with richness (alpha, gamma), occupancy (temporal, spatial, species) and turnover (temporal, spatial, species)
}
\description{
Computes occupancy statistics
}

