% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape.R
\name{get_E}
\alias{get_E}
\title{Random uniform environmental values}
\usage{
get_E(D, N, SA = NULL, XY = NULL)
}
\arguments{
\item{D}{number of environmental variables}

\item{N}{number of patches}

\item{SA}{type of autocorrelation. Default is none}
}
\description{
Random uniform environmental values
}

