% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_K_sparse}
\alias{get_K_sparse}
\title{Compute the connectivity matrix}
\usage{
get_K_sparse(XY, alpha, eps = 0.001)
}
\arguments{
\item{XY}{location of patches}

\item{alpha}{distance decay of dispersal}

\item{eps}{cutoff for sparse matrix}
}
\description{
Compute the connectivity matrix
}

