% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOutput.R
\name{plotOutput}
\alias{plotOutput}
\title{Plots an overview of the model output}
\usage{
plotOutput(out, folder = ".")
}
\arguments{
\item{out}{model output}
}
\description{
Plots an overview of the model output
}

