% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{runModel}
\alias{runModel}
\title{Main function performing the spatially explicit model}
\usage{
runModel(pars, nsteps)
}
\description{
Main function performing the spatially explicit model
}
\note{
All of the different steps are functions, provided in other scripts
}
\author{
Dominique Gravel
}

